+++ src/ActivePeerConnectionCommand.h
@@ -55,7 +55,7 @@ private:
   Time checkPoint;
   unsigned int _thresholdSpeed; // UNIT: byte/sec
   unsigned int _maxUploadSpeedLimit;
-  size_t _numNewConnection; // the number of the connection to establish.
+  unsigned int _numNewConnection; // the number of the connection to establish.
 public:
   ActivePeerConnectionCommand(int cuid,
 			      RequestGroup* requestGroup,
